requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
build==1.4.3
cachetools==7.1.1
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
cloudpathlib==0.23.0
cloudpickle==3.1.2
cohere==6.1.0
colorama==0.4.6
confection==1.3.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
cymem==2.0.13
databricks-sdk==0.106.0
Deprecated==1.3.1
distro==1.9.0
docker==7.1.0
durationpy==0.10
emoji==2.15.0
en_core_web_sm==3.8.0
fastapi==0.136.1
fastavro==1.12.2
filelock==3.25.2
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.49
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-vision==3.13.0
googleapis-common-protos==1.74.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.5.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.10.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
installer==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
langchain-core==1.2.28
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.30
llvmlite==0.47.0
lxml==6.1.0
Mako==1.3.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
narwhals==2.20.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
oauthlib==3.3.1
olefile==0.47
onnx==1.21.0
onnxruntime==1.24.4
openai==2.31.0
opencv-python==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.3.0
pikepdf==10.5.1
pillow==12.2.0
preshed==3.0.13
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.9.2
pypdfium2==5.7.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.4.20
python-magic==0.4.27
python-multipart==0.0.27
python-oxmsg==0.0.2
pytz==2026.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.5
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
skops==0.14.0
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
sqlparse==0.5.5
srsly==2.5.3
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
timm==1.0.26
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.6.2
typer==0.24.1
types-requests==2.33.0.20260503
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
unstructured==0.22.23
unstructured-client==0.43.2
unstructured.pytesseract==0.3.15
unstructured_inference==1.6.8
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.7.0
weasel==1.0.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
